test(workflows): add comprehensive unit and snapshot tests for dependency_binaries_manifest
- Add new fixtures and helpers for loading scripts and writing manifests.
- Introduce detailed tests for manifest parsing, I/O functions, argument parsing, and main entry point.
- Cover edge cases like duplicate packages, missing files, malformed TOML, and output correctness.
- Remove old monolithic test file in favor of modular test files per feature.
- Improve test coverage and stability for manifest script functionality.
