+++ .gitignore
@@ -2,4 +2,6 @@ target/
 **/*.rs.bk
 .grepai/
 bin/act
-tests/workflows/__pycache__/
+__pycache__/
+*.py[cod]
+*.so
